Add logic to header.js to open secret code form and style home page
The style given to the home page addresses the html and body elements,
in which the html envelopes the whole screen using 100% both the
width and height. On the other hand, the body element only covers
the whole width, but the height is given with flex so that the elements
inside it can grow with the page while being resized.